word cache ems page fields to byte and asm updated

+++ m_near.c
@@ -230,14 +230,14 @@ int16_t             numspritelumps;
 
 int16_t             numtextures;
 
+ 
 
-
-int16_t activetexturepages[NUM_TEXTURE_L1_CACHE_PAGES]; // always gets reset to defaults at start of frame
+uint8_t activetexturepages[NUM_TEXTURE_L1_CACHE_PAGES]; // always gets reset to defaults at start of frame
 uint8_t activenumpages[NUM_TEXTURE_L1_CACHE_PAGES];     // always gets reset to defaults at start of frame
 uint8_t textureL1LRU[NUM_TEXTURE_L1_CACHE_PAGES];
 
 
-int16_t activespritepages[NUM_SPRITE_L1_CACHE_PAGES];    // always gets reset to defaults at start of frame
+uint8_t activespritepages[NUM_SPRITE_L1_CACHE_PAGES];    // always gets reset to defaults at start of frame
 uint8_t activespritenumpages[NUM_SPRITE_L1_CACHE_PAGES]; // always gets reset to defaults at start of frame
 uint8_t spriteL1LRU[NUM_SPRITE_L1_CACHE_PAGES];
